chore: Move core/errors package
